+++ src/modules/flat/rules/setup.ts
@@ -1,5 +1,9 @@
+import { buildAddCheckRuleElement } from "./add"
 import { buildModifyFlatDCRuleElement } from "./modify"
+import { buildTreatAsRuleElement } from "./treat-as"
 
 export function setupRuleElements() {
 	game.pf2e.RuleElements.custom["fc-ModifyFlatDC"] = buildModifyFlatDCRuleElement()
+	game.pf2e.RuleElements.custom["fc-AddCheck"] = buildAddCheckRuleElement()
+	game.pf2e.RuleElements.custom["fc-TreatAs"] = buildTreatAsRuleElement()
 }